+++ data-raw/SLS.R
@@ -4,6 +4,7 @@ require(readr)
 require(dplyr)
 require(lubridate)
 require(tidyr)
+require(stringr)
 
 SLS_stations<-read_csv(file.path("data-raw", "SLS", "20mm Stations.csv"),
                        col_types=cols_only(Station="c", LatD="d", LatM="d",
@@ -25,7 +26,7 @@ SLS<-read_csv(file.path("data-raw", "SLS", "Water Info.csv"),
   mutate(Date=parse_date_time(Date, orders="%m/%d/%Y %H:%M:%S", tz="America/Los_Angeles"),
          Time=parse_date_time(Time, orders="%m/%d/%Y %H:%M:%S", tz="America/Los_Angeles"),
          Datetime=parse_date_time(if_else(is.na(Time), NA_character_, paste(year(Date), month(Date), day(Date), hour(Time), minute(Time), sep="/")),
-                                  orders="%Y/%m/%d/%H/%M"))%>%
+                                  orders="%Y/%m/%d/%H/%M", tz="America/Los_Angeles"))%>%
   mutate(Source="SLS",
          Conductivity=if_else(str_detect(Notes, "CDEC"), NA_real_, Conductivity), # Remove any conductivity values that may have come from CDEC
          Tide=recode(as.character(Tide), `4`="Flood", `3`="Low Slack", `2`="Ebb", `1`="High Slack"),