+++ NEWS.md
@@ -1,3 +1,4 @@
+* Added Stockton Dissolved Oxygen survey as an additional data source
 * Removed duplicated water quality data from suisun data
 * Added Smelt Larva Survey as an additional data source
 * Tweaked the wq function interface to force users to specify data sources, in preparation for adding more sources.